refactor(acp): extract session-usage and llm-providers clusters from agent/mod.rs
Continue the agent/mod.rs monolith split (#4843) by moving the two largest
self-contained feature clusters into dedicated, once-gated submodules:

- agent/usage.rs (unstable-session-usage): TurnUsage, SessionUsageAccumulator,
  and their tests.
- agent/providers.rs (unstable-llm-providers): ProviderSetOverride, the
  with_provider_names builder, the providers/list|set|disable handlers
  (ext_method_providers, do_list_providers, do_set_providers,
  do_disable_providers), and their tests.

Each submodule is gated once at its declaration, collapsing the per-item
are re-exported into the agent module via pub(crate) use so existing references
resolve unchanged; ext_method_providers is bumped from private to pub(crate)
for the cross-module call from the ext dispatch.

mod.rs shrinks from 3510 to 3069 lines. The gates that remain are the
irreducible ones on shared ZephAcpAgentState / SessionEntry / DrainResult
fields and inside shared method bodies, matching the issue's guidance to keep
feature-gated fields on shared structs. No public API or behavior change.

Refs #4843
